Machine Learning/Bookvault_Project/requirements.txt
# ============================================================================
# BookVault - Python Dependencies
# Tested with Python 3.13+
# ============================================================================

# ----------------------------------------------------------------------------
# Core Web Framework
# ----------------------------------------------------------------------------
streamlit>=1.52.1

# ----------------------------------------------------------------------------
# AI/ML - OpenAI Integration
# ----------------------------------------------------------------------------
openai>=2.11.0

# ----------------------------------------------------------------------------
# HTTP Client for API Requests
# ----------------------------------------------------------------------------
requests>=2.32.5

# ----------------------------------------------------------------------------
# Image Processing & OCR
# ----------------------------------------------------------------------------
Pillow>=12.0.0
pytesseract>=0.3.13

# ----------------------------------------------------------------------------
# Environment & Configuration Management
# ----------------------------------------------------------------------------
python-dotenv>=1.2.1

# ----------------------------------------------------------------------------
# Async Support (for future async operations)
# ----------------------------------------------------------------------------
aiohttp>=3.13.2

# ----------------------------------------------------------------------------
# Fuzzy Matching for Search Intelligence
# Used for typo-tolerant search and "Did you mean?" suggestions
# ----------------------------------------------------------------------------
rapidfuzz>=3.6.1

# ----------------------------------------------------------------------------
# Testing & Quality Assurance
# ----------------------------------------------------------------------------
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=1.3.0
